gifgenerator/requirements.txt
appdirs==1.4.4
appnope==0.1.3
asgiref==3.3.4
attrs==21.4.0
backcall==0.2.0
bleach==5.0.1
branca==0.4.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.10
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
cycler==0.11.0
dbbackup-tools==0.1.5
decorator==4.4.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==2.2.16
django-allauth==0.44.0
django-appconf==1.0.4
django-auditlog==2.1.1
django-cors-headers==3.4.0
django-crontab==0.7.1
django-dbbackup==3.3.0
django-extensions==3.1.3
django-maintenance-mode==0.16.2
django-postgres-extensions==0.9.3
django-rest-swagger==2.1.2
djangorestframework==3.10.3
drf-api-logger==1.1.10
drf-openapi==1.3.0
drf-yasg==1.21.4
et-xmlfile==1.1.0
factory-boy==3.2.1
Faker==13.12.0
folium==0.12.1
fonttools==4.29.1
geographiclib==1.52
geopy==2.2.0
gunicorn==20.1.0
idna==2.10
imageio==2.24.0
imageio-ffmpeg==0.4.8
inflection==0.5.1
iniconfig==1.1.1
ipython==7.25.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.1
kiwisolver==1.3.2
Mako==1.2.0
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
moviepy==1.0.3
numpy==1.22.2
oauthlib==3.1.0
openapi-codec==1.3.2
opencv-python==4.5.5.62
opencv-python-headless==4.5.5.62
openpyxl==3.0.9
packaging==21.3
pandas==1.3.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pluggy==1.0.0
proglog==0.1.10
prompt-toolkit==3.0.29
psycopg2==2.9.3
psycopg2-binary==2.9.5
ptyprocess==0.7.0
py==1.11.0
pycparser==2.20
pygame==2.1.2
pygame-gui==0.6.8
pygame-menu==4.3.6
Pygments==2.12.0
PyJWT==2.1.0
pyparsing==3.0.7
pyperclip==1.8.2
PySimpleGUI==4.56.0
pytest==7.1.2
pytest-django==4.5.2
pytest-factoryboy==2.3.1
python-dateutil==2.8.2
python-decouple==3.5
python-i18n==0.3.9
python3-openid==3.2.0
pytz==2021.1
PyYAML==6.0
qrcode==7.3.1
reportlab==3.5.67
requests==2.27.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sentry-sdk==1.5.4
simplejson==3.18.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
tomli==2.0.1
tqdm==4.64.1
traitlets==5.2.1.post0
typing-extensions==3.10.0.0
Unidecode==1.3.4
uritemplate==4.1.1
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
whitenoise==5.3.0
xlrd==2.0.1
XlsxWriter==3.0.2
